Setup Clicks, Deletion, Animation
Removed 2-way-view in favor of more simplistic items. Based on research most people have used the ViewHolder for onItemClick listener support, sometimes I've seen a ViewHolder implement View.OnClickListener, however that didn't work for our example so our root view has an itemclicklistener set.

On top of that, basic animations were added via a top starred github library.

The delete is hooked up to actually remove the item from the list. However it seems glitchy in comparison.

It should be noted that it will be left up to the user to determine UI/UX aspects such as removal. The Inbox App by Google uses a swipe to remove items, where the Gmail App by Google uses a swipe to 'queue' an item for dismisal, but not dismissing it until the user makes another interaction such as scrolling or clicking another email.

A 'closeAllItems()' method was added in our Managers, as well as a minor fix that was needed by adding 'shown layouts' correctly in our SwipeListener.